standardize grouping.
